requirements.txt
python-telegram-bot
requests
pytz
flask
openai